Fix triggering of path change
Sometimes the runtime will delay a process longer than is requested sleep. In
this case, the transition may have already started the `Enter` phase, which
causes `start` to be reset. Becuase the `start` value is different from the
value expected by the delayed process, it fizzles.

This fix keeps the value in `start` constant and adjusts the animation math in
the `Enter` phase to account for `start` meaning the start of the animation
instead of the start of the phase.